move e2e test helpers to src/e2e/helpers directory
