+++ src/e2e/helpers/clickBullet.ts
