+++ src/e2e/helpers/clickThought.ts
