+++ src/e2e/helpers/clickWithOffset.ts
