+++ src/e2e/helpers/getEditingText.ts
