+++ src/e2e/helpers/getLexemeFromDB.ts
