+++ src/e2e/helpers/initPage.ts
