+++ src/e2e/helpers/mobile/clickThought.ts
