+++ src/e2e/helpers/mobile/clickWithOffset.ts
