+++ src/e2e/helpers/mobile/gestures/newThoughtGesture.ts
