+++ src/e2e/helpers/mobile/getEditable.ts
