+++ src/e2e/helpers/mobile/getEditingText.ts
