+++ src/e2e/helpers/mobile/getNativeElementRect.ts
