+++ src/e2e/helpers/mobile/hideKeyboardByTappingDone.ts
