+++ src/e2e/helpers/mobile/paste.ts
