+++ src/e2e/helpers/mobile/tapReturnKey.ts
