+++ src/e2e/helpers/paste.ts
