+++ src/e2e/helpers/skipTutorialScreen.ts
