+++ src/e2e/helpers/waitForElementBecomeHidden.ts
