+++ src/e2e/helpers/waitForThoughtExistInDb.ts
